+++ OneBlink.SDK/models/FormSubmissionApproval.cs
@@ -56,5 +56,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public List<FormSubmissionApprovalNote> additionalNotes
+        {
+            get; set;
+        }
     }
 }