+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+-   `additionalNotes` to `FormSubmissionApproval`
+
 ## 3.0.6 (2022-06-17)
 
 ### Added